third_party/blink/web_tests/platform/mac/external/wpt/css/css-fonts/matching/font-unicode-PUA-expected.txt

This is a testharness.js-based test.
[FAIL] PUA character U+F000 is rendered with a non-generic font.
  assert_equals: expected 22 but got 15.890625
Harness: the test ran to completion.
